+++ manticore/core/cpu/x86.py
@@ -12,7 +12,7 @@ from .abstractcpu import (
 
 
 from ..smtlib import Operators, BitVec, Bool, BitVecConstant, operator, visitors
-from ..memory import ConcretizeMemory
+from manticore.native.memory import ConcretizeMemory
 from ...utils.helpers import issymbolic
 from functools import reduce
