+++ manticore/core/plugin.py
@@ -1,8 +1,5 @@
 import logging
 from contextlib import contextmanager
-from typing import Optional
-
-from capstone import CS_GRP_JUMP
 
 from ..utils.helpers import issymbolic
 
@@ -243,36 +240,6 @@ class Visited(Plugin):
                 for m in executor_visited:
                     f.write(f"0x{m:016x}\n")
         logger.info('Coverage: %d different instructions executed', len(executor_visited))
-
-
-class ConcreteTraceFollower(Plugin):
-    def __init__(self, source=None):
-        '''
-        :param iterable source: Iterator producing instruction pointers to be followed
-        '''
-        super().__init__()
-        self.source = source
-
-    def will_start_run_callback(self, state):
-        self.saved_flags = None
-
-    def will_execute_instruction_callback(self, state, pc, instruction):
-        if not instruction.group(CS_GRP_JUMP):
-            self.saved_flags = None
-            return
-
-        # Likely unconditional
-        if not instruction.regs_read:
-            self.saved_flags = None
-            return
-
-        self.saved_flags = state.cpu.RFLAGS
-        state.cpu.RFLAGS = state.new_symbolic_value(state.cpu.address_bit_size)
-
-    def did_execute_instruction_callback(self, state, pc, target_pc, instruction):
-        # Should direct execution via trace
-        if self.saved_flags:
-            state.cpu.RFLAGS = self.saved_flags
 
 
 # TODO document all callbacks
