+++ manticore/native/mappings.py
