+++ manticore/platforms/decree.py
@@ -3,7 +3,7 @@ from . import cgcrandom
 # TODO use cpu factory
 from ..core.cpu.x86 import I386Cpu
 from ..core.cpu.abstractcpu import Interruption, ConcretizeRegister, ConcretizeArgument
-from ..core.memory import SMemory32, Memory32
+from manticore.native.memory import SMemory32, Memory32
 from ..core.smtlib import *
 from ..core.executor import TerminateState
 from ..utils.helpers import issymbolic
