+++ manticore/platforms/linux.py
@@ -18,7 +18,7 @@ from elftools.elf.descriptions import describe_symbol_type
 from ..core.smtlib import Expression
 from ..core.cpu.abstractcpu import Interruption, Syscall, ConcretizeArgument
 from ..core.cpu.cpufactory import CpuFactory
-from ..core.memory import SMemory32, SMemory64, Memory32, Memory64
+from manticore.native.memory import SMemory32, SMemory64, Memory32, Memory64
 from ..core.smtlib import Operators, ConstraintSet, solver
 from ..exceptions import SolverException
 from ..core.cpu.arm import *
