+++ tests/test_abi.py
@@ -2,22 +2,12 @@
 # -*- coding: utf-8 -*-
 
 import unittest
-import sys
-import shutil
-import tempfile
-import os
-import hashlib
-import subprocess
-import collections
-import time
 
-from manticore import Manticore, issymbolic
 from manticore import variadic
-from manticore.core.smtlib import BitVecVariable
 from manticore.core.cpu.abstractcpu import ConcretizeArgument, ConcretizeRegister, ConcretizeMemory
 from manticore.core.cpu.arm import Armv7Cpu, Armv7LinuxSyscallAbi, Armv7CdeclAbi
 from manticore.core.cpu.x86 import I386Cpu, AMD64Cpu, I386LinuxSyscallAbi, I386StdcallAbi, I386CdeclAbi, AMD64LinuxSyscallAbi, SystemVAbi
-from manticore.core.memory import SMemory32, Memory32, SMemory64
+from native.memory import SMemory32, SMemory64
 from manticore.core.smtlib import ConstraintSet, Operators
 
 class ABITest(unittest.TestCase):
