+++ tests/test_cpu_automatic.py
@@ -1,8 +1,7 @@
 
 import unittest
 from manticore.core.cpu.x86 import *
-from manticore.core.smtlib import *
-from manticore.core.memory import *
+from native.memory import *
 
 class CPUTest(unittest.TestCase):
     _multiprocess_can_split_ = True
