+++ tests/test_cpu_manual.py
@@ -3,7 +3,7 @@ import struct
 import unittest
 from manticore.core.cpu.x86 import *
 from manticore.core.smtlib import Operators
-from manticore.core.memory import *
+from native.memory import *
 from manticore.core.smtlib import BitVecOr
 from tests import mockmem
 from functools import reduce
