+++ tests/test_slam_regre.py
@@ -1,7 +1,7 @@
 import unittest
 from manticore.core.cpu.x86 import *
 from manticore.core.smtlib import Operators
-from manticore.core.memory import *
+from native.memory import *
 
 
 class CPUTest(unittest.TestCase):
